Interchain Process Test 분리
